Websockets logic moved to submodules
