+++ src/components/projects/projectId/heuristics/HeuristicsHeader.tsx
@@ -7,7 +7,7 @@ import { selectProjectId } from '@/src/reduxStore/states/project';
 import style from '@/src/styles/components/projects/projectId/heuristics/heuristics.module.css';
 import { Heuristic, HeuristicsHeaderProps } from '@/src/types/components/projects/projectId/heuristics/heuristics';
 import { LabelingTask } from '@/src/types/components/projects/projectId/settings/labeling-tasks';
-import { CurrentPage } from '@/src/types/shared/general';
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { ModalEnum } from '@/src/types/shared/modal';
 import { Status } from '@/src/types/shared/statuses';
 import { ACTIONS_DROPDOWN_OPTIONS, NEW_HEURISTICS, checkSelectedHeuristics, postProcessCurrentWeakSupervisionRun } from '@/src/util/components/projects/projectId/heuristics/heuristics-helper';
@@ -21,7 +21,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import LastWeakSupervisionModal from './modals/LastWeakSupervisionModal';
 import DeleteHeuristicsModal from './DeleteHeuristicsModal';
 import Dropdown2 from '@/submodules/react-components/components/Dropdown2';
-import { useWebsocket } from '@/src/services/base/web-sockets/useWebsocket';
+import { useWebsocket } from '@/submodules/react-components/hooks/web-socket/useWebsocket';
 import { createTask, getWeakSupervisionRun, setAllHeuristics } from "@/src/services/base/heuristic";
 import { initZeroShot } from "@/src/services/base/zero-shot";
 import { initWeakSupervision } from "@/src/services/base/weak-supervision";
