+++ src/components/projects/projectId/heuristics/heuristicId/crowd-labeler/CrowdLabeler.tsx
@@ -1,7 +1,7 @@
 import { selectHeuristic, setActiveHeuristics } from "@/src/reduxStore/states/pages/heuristics";
 import { selectLabelingTasksAll, setLabelingTasksAll } from "@/src/reduxStore/states/pages/settings";
 import { selectProjectId } from "@/src/reduxStore/states/project";
-import { CurrentPage } from "@/src/types/shared/general";
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { postProcessLabelingTasks, postProcessLabelingTasksSchema } from "@/src/util/components/projects/projectId/settings/labeling-tasks-helper";
 import { useRouter } from "next/router";
 import { useCallback, useEffect } from "react";
@@ -17,7 +17,7 @@ import { selectDataSlicesAll, setDataSlices } from "@/src/reduxStore/states/page
 import { getAllComments } from "@/src/services/base/comment";
 import { CommentDataManager } from "@/src/util/classes/comments";
 import { CommentType } from "@/src/types/shared/comments";
-import { useWebsocket } from "@/src/services/base/web-sockets/useWebsocket";
+import { useWebsocket } from "@/submodules/react-components/hooks/web-socket/useWebsocket";
 import { getDataSlices } from "@/src/services/base/dataSlices";
 import { getLabelingTasksByProjectId } from "@/src/services/base/project";
 import { getAccessLink, getHeuristicByHeuristicId } from "@/src/services/base/heuristic";
