+++ src/components/projects/projectId/settings/ProjectSnapshotExportModal.tsx
@@ -6,10 +6,10 @@ import { selectEmbeddings } from "@/src/reduxStore/states/pages/settings";
 import { selectProjectId } from "@/src/reduxStore/states/project";
 import { getProjectSize, prepareProjectExport } from "@/src/services/base/project-setting";
 import { downloadFile } from "@/src/services/base/s3-service";
-import { useWebsocket } from "@/src/services/base/web-sockets/useWebsocket";
+import { useWebsocket } from "@/submodules/react-components/hooks/web-socket/useWebsocket";
 import { getLastProjectExportCredentials } from '@/src/services/base/project';
 import { DownloadState, ProjectSize } from "@/src/types/components/projects/projectId/settings/project-export";
-import { CurrentPage, CurrentPageSubKey } from "@/src/types/shared/general";
+import { CurrentPage, CurrentPageSubKey } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { ModalEnum } from "@/src/types/shared/modal";
 import { postProcessingFormGroups } from "@/src/util/components/projects/projectId/settings/project-export-helper";
 import { TOOLTIPS_DICT } from "@/src/util/tooltip-constants";
