+++ src/components/shared/admin-messages/AdminMessages.tsx
@@ -1,6 +1,6 @@
 import { selectCurrentPage } from "@/src/reduxStore/states/general";
 import { AdminMessageLevel, AdminMessagesProps } from "@/src/types/shared/admin-messages";
-import { CurrentPage } from "@/src/types/shared/general";
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { IconAlertCircle, IconInfoSquare, IconPoint, IconX } from "@tabler/icons-react";
 import { useSelector } from "react-redux";
