+++ src/components/shared/comments/Comments.tsx
@@ -5,12 +5,12 @@ import { IconNotes } from "@tabler/icons-react";
 import { useDispatch, useSelector } from "react-redux";
 import { useCallback, useState } from "react";
 import { selectAllUsers, setAllUsers, setComments } from "@/src/reduxStore/states/general";
-import { CurrentPage, CurrentPageSubKey } from "@/src/types/shared/general";
+import { CurrentPage, CurrentPageSubKey } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { CommentDataManager } from "@/src/util/classes/comments";
 import { CommentRequest, CommentType } from "@/src/types/shared/comments";
 import { commentRequestToKey } from "@/src/util/shared/comments-helper";
 import { selectProjectId } from "@/src/reduxStore/states/project";
-import { useWebsocket } from "@/src/services/base/web-sockets/useWebsocket";
+import { useWebsocket } from "@/submodules/react-components/hooks/web-socket/useWebsocket";
 import { getAllComments } from "@/src/services/base/comment";
 import { getOrganizationUsers } from "@/src/services/base/organization";
