+++ src/pages/projects/[projectId]/attributes/[attributeId]/index.tsx
@@ -1,6 +1,6 @@
 import AttributeCalculation from "@/src/components/projects/projectId/attributes/attributeId/AttributeCalculations";
 import { setCurrentPage, setDisplayIconComments } from "@/src/reduxStore/states/general";
-import { CurrentPage } from "@/src/types/shared/general";
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
