+++ src/pages/projects/[projectId]/settings/index.tsx
@@ -1,6 +1,6 @@
 import ProjectSettings from "@/src/components/projects/projectId/settings/ProjectSettings";
 import { setCurrentPage, setDisplayIconComments } from "@/src/reduxStore/states/general";
-import { CurrentPage } from "@/src/types/shared/general";
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
