+++ src/pages/projects/new/index.tsx
@@ -1,6 +1,6 @@
 import NewProject from "@/src/components/projects/new/NewProject"
 import { setCurrentPage, setDisplayIconComments } from "@/src/reduxStore/states/general"
-import { CurrentPage } from "@/src/types/shared/general"
+import { CurrentPage } from "@/submodules/react-components/hooks/web-socket/web-sockets-helper";
 import { useEffect } from "react"
 import { useDispatch } from "react-redux"
