+++ src/types/shared/general.ts
@@ -1,44 +1,5 @@
 import { UserType } from "../components/projects/projectId/labeling/labeling-main-component";
 
-export enum CurrentPage {
-    PROJECTS = 'PROJECTS',
-    LOOKUP_LISTS_OVERVIEW = 'LOOKUP_LISTS_OVERVIEW',
-    NEW_PROJECT = 'NEW_PROJECT',
-    PROJECT_OVERVIEW = 'PROJECT_OVERVIEW',
-    DATA_BROWSER = 'DATA_BROWSER',
-    LABELING = 'LABELING',
-    HEURISTICS = 'HEURISTICS',
-    ADMIN_PAGE = 'ADMIN_PAGE',
-    USERS = 'USERS',
-    UPLOAD_RECORDS = 'UPLOAD_RECORDS',
-    PROJECT_SETTINGS = "PROJECT_SETTINGS",
-    MODELS_DOWNLOAD = "MODELS_DOWNLOAD",
-    ATTRIBUTE_CALCULATION = "ATTRIBUTE_CALCULATION",
-    LOOKUP_LISTS_DETAILS = "LOOKUP_LISTS_DETAILS",
-    MODEL_CALLBACKS = "MODEL_CALLBACKS",
-    LABELING_FUNCTION = "LABELING_FUNCTION",
-    ACTIVE_LEARNING = "ACTIVE_LEARNING",
-    ZERO_SHOT = "ZERO_SHOT",
-    CROWD_LABELER = "CROWD_LABELER",
-    RECORD_IDE = "RECORD_IDE",
-    EDIT_RECORDS = "EDIT_RECORDS",
-    NOTIFICATION_CENTER = "NOTIFICATION_CENTER",
-    EXPORT = "EXPORT",
-    COMMENTS = "COMMENTS",
-    BRICKS_INTEGRATOR = "BRICKS_INTEGRATOR",
-    CONFIG = "CONFIG",
-}
-
-export enum CurrentPageSubKey {
-    EMBEDDINGS = 'EMBEDDING',
-    NONE = 'NONE',
-    BUTTONS_CONTAINER = 'BUTTONS_CONTAINER',
-    VARIABLE_SELECTION = 'VARIABLE_SELECTION',
-    SNAPSHOT_EXPORT = 'SNAPSHOT_EXPORT',
-    GLOBAL = 'GLOBAL',
-    FILE_UPLOAD = 'FILE_UPLOAD',
-}
-
 export type User = {
     id: string;
     firstName: string;
